Add files via upload
Example IR transmittance data for solutions of benzyl alcohol, toluene and a mixture of the two in an unknown proportion. The file "determining_composition_of_mixture_of_benzyl_alcohol_and_toluene.py" is the code that allows the composition of the mixture of toluene and benzyl alcohol to be determined from IR data. The file "benzyl_alcohol_toluene_mixture_functions_module.py" is a module of functions used in the determination of the composition of the mixture